fixup windows tests
